aws_credential_types

Module credential_fn

Types that allow a credentials provider to be created from a closure

Structs

ProvideCredentialsFn

A ProvideCredentials implemented by a closure.

Functions

provide_credentials_fn

Returns a new credentials provider built with the given closure. This allows you to create an ProvideCredentials implementation from an async block that returns a crate::provider::Result.