Module aws_credential_types::token_fn

Types that allow an access token provider to be created from a closure

Structs

ProvideTokenFn

A ProvideToken implemented by a closure.

Functions

provide_token_fn

Returns a new token provider built with the given closure. This allows you to create an ProvideToken implementation from an async block that returns a crate::provider::token::Result.