Crate aws_credential_types

aws-credential-types provides types concerned with AWS SDK credentials including:

Traits for credentials providers and for credentials caching

An opaque struct representing credentials

Concrete implementations of credentials caching

Modules

attributes

Types representing specific pieces of data contained within credentials or within token

credential_fn

Types that allow a credentials provider to be created from a closure

provider

AWS SDK Credentials

token_fn

Types that allow an access token provider to be created from a closure

Structs

Credentials

AWS SDK Credentials

CredentialsBuilder

Builder for Credentials

Type Aliases

Token

AWS Access Token